feat(observability): add missing dashboards and fix rust build
- Add `ohc_agent_task_queue_depth` metric to `src/server/orchestration/kairos.rs`
- Add missing `chrono` dependency and `FundingOpportunity` struct in `src/server/services/capital/funding_engine.rs`
- Remove non-existent `jobs.rs` from `src/server/api/agents/BUILD.bazel`
- Create and provision `agent_api_errors_dashboard.json` and `kairos_dashboard.json`
- Update `llm_provider_health.json` to match specifications
- Resolve struct re-declaration and type conflicts from origin/main rebase

+++ src/server/api/capital/funding_api.rs
@@ -38,7 +38,7 @@ impl FundingService for FundingApi {
             id: o.id,
             tenant_id: o.tenant_id,
             grant_name: o.grant_name,
-            amount: o.amount,
+            amount: o.amount as f64,
             draft_proposal_text: o.draft_proposal_text.unwrap_or_default(),
             status: o.status,
             deadline: o.deadline.map(|d| d.to_rfc3339()).unwrap_or_default(),
